refactor(refactor-logic): 승패 결정 로직 이동
승패를 결정하는 로직을 전용 클래스를 만들어서 분리했습니다.
